More cleanups to level provider hierarchy, move more garbage out of BaseLevelProvider
